Updated routes.js and mk_permit.js with Tom's code for extracting the risk data as JSON from atcouncil.org's page.
Modified f.set_coordinates() to perform the task of obtaining the wind data using Tom's code if the lat and lon come through to that point.

Added a loading, checkbox styled and red crossbox style images to public/data to help signal the status of obtaining Address and Wind data since the process takes a couple seconds to perform.

The wind data is placed into g.system.location.wind so that the wind data will correspond with the location that it belongs to rather than being a separate entity (eg: g.system.wind)